\name{asRisk-class}
\docType{class}
\alias{asRisk-class}

\title{Aymptotic risk}
\description{Class of asymptotic risks.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{type}:}{Object of class \code{"character"}.}
  }
}
\section{Extends}{
Class \code{"RiskType"}, directly.
}
\section{Methods}{
No methods defined with class "asRisk" in the signature.
}
\references{
  Rieder, H. (1994) \emph{Robust Asymptotic Statistics}. New York: Springer.

  Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for
  General Loss Functions. Statistics & Decisions (submitted).

  Kohl, M. (2005) \emph{Numerical Contributions to the Asymptotic Theory of Robustness}. 
  Bayreuth: Dissertation.
}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{RiskType-class}}}
%\examples{}
\concept{asymptotic risk}
\concept{risk}
\keyword{classes}
