\name{fiRisk-class}
\docType{class}
\alias{fiRisk-class}

\title{Finite-sample risk}
\description{Class of finite-sample risks.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{type}:}{Object of class \code{"character"}.}
  }
}
\section{Extends}{
Class \code{"RiskType"}, directly.
}
\section{Methods}{
No methods defined with class "fiRisk" in the signature.
}
\references{
  Ruckdeschel, P. and Kohl, M. (2005) How to approximate 
  the finite sample risk of M-estimators. 
}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{RiskType-class}}}
%\examples{}
\concept{finite-sample risk}
\concept{risk}
\keyword{classes}
