\name{MatrixorFunction-class}
\docType{class}
\alias{MatrixorFunction-class}

\title{MatrixorFunction-class}
\description{Optional object of class \code{"matrix"} of of class \code{"function"}.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
%\section{Methods}{}
%\references{}
\author{Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de}}
%\note{}
\seealso{\code{\link[methods]{matrix-class}}, \code{\link[methods:BasicClasses]{function-class}}}
\keyword{classes}
