\name{NonSymmetric}
\alias{NonSymmetric}

\title{Generating function for NonSymmetric-class}
\description{Generates an object of class \code{"NonSymmetric"}.}
\usage{NonSymmetric()}
%\details{}
\value{Object of class \code{"NonSymmetric"}}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{NonSymmetric-class}}, \code{\link{FunctionSymmetry-class}}}
\examples{
NonSymmetric()

## The function is currently defined as
function(){ new("NonSymmetric") }
}
\concept{symmetry}
\keyword{math}
